server/requirements.txt
alembic==1.16.5
blinker==1.9.0
click==8.2.1
Flask==3.1.2
greenlet==3.2.4
inflect==7.5.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
measured==0.12.2
more-itertools==10.8.0
numpy==2.3.3
pandas==2.3.2
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
SQLAlchemy==2.0.43
SQLAlchemy-Utils==0.42.0
typeguard==4.4.4
typing_extensions==4.15.0
tzdata==2025.2
Werkzeug==3.1.3
